requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
PyQt6>=6.4.0
openai>=1.0.0
mistralai>=1.0.0
pyperclip>=1.8.2
soundfile>=0.12.1
pydub>=0.25.1
python-dotenv>=1.0.0
